Add Delete ,update and mytravel view funtions,templates and urls

+++ userapp/views.py
@@ -21,6 +21,44 @@ today = date.today()
 
 from .forms import BusForm, SeatsForm,CreateUserForm,BookForm
 
+def registeruser(request):
+    if request.method == 'POST':
+        form = CreateUserForm(request.POST)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Account Created Successfully!. Check out our Email later :)')
+
+            return redirect('loginpage')
+    else:
+        form = CreateUserForm
+    context = {
+            
+            'form':form,
+                        }
+
+    return render(request,'registration/register.html',context)
+
+def loginpage(request):
+    if request.user.is_authenticated:
+
+            return redirect('index')
+    else:
+        if request.method == 'POST':
+            username = request.POST.get('username')
+            password =request.POST.get('password')
+
+            user = authenticate(request, username=username, password=password)
+
+            if user is not None:
+                login(request, user)
+                return redirect('index')
+            else:
+                messages.info(request, 'Username or password is incorrect')
+
+      
+    return render(request,'registration/login.html')
+
+
 @login_required(login_url='loginpage')
 def index(request):
     buses = Bus.objects.all()
@@ -62,45 +100,12 @@ def confirm_booking(request):
 
 
 def mytravels(request):
-    mytravel=Book.objects.last()
+    current_user=request.user.email
+
+    mytravel=Book.show_bookings(current_user)
+
     return render(request,'main/mytravel.html',{'mytravel':mytravel})
 
-def registeruser(request):
-    if request.method == 'POST':
-        form = CreateUserForm(request.POST)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Account Created Successfully!. Check out our Email later :)')
-
-            return redirect('loginpage')
-    else:
-        form = CreateUserForm
-    context = {
-            
-            'form':form,
-                        }
-
-    return render(request,'registration/register.html',context)
-
-def loginpage(request):
-    if request.user.is_authenticated:
-
-            return redirect('index')
-    else:
-        if request.method == 'POST':
-            username = request.POST.get('username')
-            password =request.POST.get('password')
-
-            user = authenticate(request, username=username, password=password)
-
-            if user is not None:
-                login(request, user)
-                return redirect('index')
-            else:
-                messages.info(request, 'Username or password is incorrect')
-
-      
-    return render(request,'registration/login.html')
 
 
 def logout_page(request):
@@ -108,24 +113,30 @@ def logout_page(request):
     return redirect('loginpage')
 
 def checkout(request):
-    books=Book.objects.all()
+    books=Book.objects.last()
     return render(request,'main/checkout.html',{'books':books})
 
 def update(request,pk):
-    book = Book.objects.get(pk=pk)
+    book = Book.objects.get(id=pk)
+    form=SeatsForm(instance=book)
     if request.method == 'POST':
-        form = BookForm(request.POST, instance=book)
+        form = SeatsForm(request.POST, instance=book)
         if form.is_valid():
             form.save()
-            url = reverse('index', kwargs={'pk': pk})
-            return render(request, 'index.html', {'url': url})
-        else:
-            form = BookForm(instance=book)
-    else:
-        form = BookForm(instance=book)
-    return render(request, 'update.html', {'form':form, 'book':book})
+            return redirect('mytravels')
+        
+        
+    return render(request, 'update.html', {'form':form})
+
 
 def delete_booking(request, pk):
-    book = Book.objects.get(pk=pk)
-    book.delete()
-    return render(request, 'index.html')
+	book = Book.objects.get(id=pk)
+	if request.method == "POST":
+		book.delete()
+		return redirect('mytravels')
+
+	context = {'book':book}
+	return render(request, 'delete.html', context)
+
+
+
